try de ruta, usuario_ruta y usuario

+++ api/ruta.py
@@ -16,29 +16,42 @@ def getruta():
 
 @ruta_ruta.route("/saveruta", methods=["POST"])
 def saveruta():
-    id_ruta = request.json['id_ruta']
-    punto_inicio= request.json['punto_inicio']
-    punto_final= request.json['punto_final']
-    new_ruta= ruta(id_ruta,punto_inicio,punto_final)
-    db.session.add(new_ruta)
-    db.session.commit()
-    return "Datos guardados con exitos"
+    try:
+        id_ruta = request.json['id_ruta']
+        punto_inicio= request.json['punto_inicio']
+        punto_final= request.json['punto_final']
+        new_ruta= ruta(id_ruta,punto_inicio,punto_final)
+        db.session.add(new_ruta)
+        db.session.commit()
+        return "Datos guardados con exitos"
+    except Exception as e:
+        return f"Hubo un problema{str(e)}"
 
-@ruta_ruta.route("/updateruta", methods=["PUT"])
-def updateruta():
-    id_ruta = request.json['id_ruta']
-    punto_inicio= request.json['punto_inicio']
-    punto_final= request.json['punto_final']
-    nruta = ruta.query.get(id_ruta) #Select * from Cliente where id = id
-    nruta.punto_inicio=punto_inicio
-    nruta.punto_final=punto_final  
-    db.session.commit()
-    return "Datos Actualizado con exitos"
+@ruta_ruta.route("/updateruta/<id>", methods=["PUT"])
+def updateruta(id):
+    try:
+        id_ruta = ruta.query.get(id)
+        if not id_ruta:
+            return "El usuario no esta registrado"
+        punto_inicio= request.json['punto_inicio']
+        punto_final= request.json['punto_final']
+        nruta = ruta.query.get(id_ruta) #Select * from Cliente where id = id
+        nruta.punto_inicio=punto_inicio
+        nruta.punto_final=punto_final  
+        db.session.commit()
+        return "Datos Actualizado con exitos"
+    except Exception as e:
+        return f"Hubo un error{str(e)}"
 
 @ruta_ruta.route("/deleteruta/<id>", methods=["DELETE"])
 def deleteruta(id):
-    id_ruta = request.json['id_ruta']
-    rutax = ruta.query.get(id_ruta)
-    db.session.delete(rutax)
-    db.session.commit()
+    try:
+        id_ruta = ruta.query.get(id)
+        if not id_ruta:
+            return "El usuario no esta registrado"
+        rutax = ruta.query.get(id_ruta)
+        db.session.delete(rutax)
+        db.session.commit()
+    except Exception as e:
+        return f"Hubo un error {str(e)}"
